:recycle: refactor: rework folders architecture
