+++ tests/e2e/CommonFunctions/goToBoardGithubIssues.ts
