+++ tests/e2e/CommonFunctions/goToGraphPage.ts
