+++ tests/e2e/CommonFunctions/openAndCloseAccordions.ts
