+++ tests/e2e/ContactPage/clickCommunityButton.ts
