+++ tests/e2e/GettingStartedPage/clickOnDDDLink.ts
