+++ tests/e2e/GettingStartedPage/clickOnFirstDocTopic.ts
