+++ tests/e2e/GraphPage/openGraphLegend.ts
