+++ tests/e2e/GraphPage/searchForKartPattern.ts
